Revert "File restructure"
