+++ curriculum-ui/src/action-mappers/categories-getall-action-mappers.ts
@@ -1,5 +1,5 @@
 import { Dispatch } from "redux";
-import { FetchAllCategories } from '../remote/category-remote'
+import { FetchAllCategories } from '../remote/Category'
 import { Category } from "../models/Category";
