+++ curriculum-ui/src/action-mappers/get-all-visualizations-action-mapper.ts
@@ -1,5 +1,5 @@
 import { Dispatch } from "redux"
-import { categoriesTypes } from "./get-all-categories-action-mapper"
+import { categoriesTypes } from "./categories-getall-action-mappers"
 import { getAllVisualizations } from "../remote/visualization-remote"
