+++ curriculum-ui/src/components/create-curriculum-form-component/CreateCurriculumFormContainer.tsx
