+++ curriculum-ui/src/components/view-all-categories-component/ViewAllCategoriesContainer.ts
@@ -1,6 +1,6 @@
 import { IState } from "../../reducers"
 import { ViewAllCategoriesComponent } from "./ViewAllCategoriesComponent"
-import { getAllCategoriesActionMapper } from "../../action-mappers/get-all-categories-action-mapper"
+import { getAllCategoriesActionMapper } from "../../action-mappers/categories-getall-action-mappers"
 import { connect } from "react-redux"
 
 const mapStateToProps = (state:IState) => {
