+++ curriculum-ui/src/components/view-all-curriculum-component/ViewAllCurriculumContainer.tsx
@@ -1,7 +1,7 @@
 import { IState } from "../../reducers"
 import { connect } from "react-redux"
 import { ViewAllCurriculumComponent } from "./ViewAllCurriculumComponent"
-import { viewCurriculumListActionMapper } from '../../action-mappers/get-all-curriculum-action-mapper';
+import { viewCurriculumListActionMapper } from '../../action-mappers/view-curriculum-list-action-mapper';
 
 const mapStateToProps = (state:IState) => {
   return({
