+++ curriculum-ui/src/components/visualization-components/create-visualization-component/CreateVisualizationContainer.tsx
@@ -1,7 +1,7 @@
 import { IState } from "../../../reducers";
 import { connect } from "react-redux";
 import { CreateVisualizationComponent } from "./CreateVisualizationComponent";
-import {createVisualizationActionMapper} from "../../../action-mappers/create-visualization-action-mapper"
+import {createVisualizationActionMapper} from "../../../action-mappers/create-visualization-Action-mapper"
 
 const mapStateToProps = (state:IState) => {
   return({
