+++ curriculum-ui/src/reducers/curriculum-reducer.ts
