+++ curriculum-ui/src/remote/CurriculumCurriculaRemote.ts
@@ -1,4 +1,4 @@
-import { curriculaClient } from "./curricula-client";
+import { curriculaClient } from "./CurriculaClient";
 import { Curriculum } from "../models/Curriculum";
